[chore] #98 video data

+++ domain/video/src/main/java/com/record/video/model/VideoData.kt
@@ -1,14 +1,14 @@
 package com.record.video.model
 
 data class VideoData(
-    val id: Int,
+    val id: Long,
     val isBookmark: Boolean,
     val bookmarkCount: Int,
     val content: String,
     val videoUrl: String,
     val previewUrl: String,
     val location: String,
-    val uploaderId: Int,
+    val uploaderId: Long,
     val nickname: String,
     val isMine: Boolean,
 )
